Add code to handle misc window.open calls in the widget
sometimes a user will click a button or a link and go to a new tab or
window. This is usually for going to the bank's website or viewing
disclosures.

This handles those kind of requests so the webview isn't over writing
the widget session with the new url, but opening the user agent to the
url.